[Breaking change] fix: removed GamesCollection class - Replaced with a Set<BoardGame, GameStats> attribute for the User entity - Also, fixed `addFriends` feature by replacing @EqualsAndHashCode Lombok annotation with recommended Hibernate functions, preventing users from inserting friends that already formed a friendship relation with them
