--- tindergames/src/main/java/org/teamcifo/tindergames/gamesCollectionEntity/GamesCollectionRepository.java
@@ -1,6 +1,0 @@
-package org.teamcifo.tindergames.gamesCollectionEntity;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface GamesCollectionRepository extends CrudRepository<GamesCollection, String> {
-}
