+++ tindergames/src/main/java/org/teamcifo/tindergames/userEntity/User.java
@@ -1,23 +1,23 @@
 package org.teamcifo.tindergames.userEntity;
 
-import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+import org.hibernate.Hibernate;
 import org.hibernate.annotations.GenericGenerator;
 import org.teamcifo.tindergames.boardGameEntity.BoardGame;
 import org.teamcifo.tindergames.gamePlayEntity.Gameplay;
-import org.teamcifo.tindergames.gamesCollectionEntity.GamesCollection;
+import org.teamcifo.tindergames.gamesCollectionEntity.GameStats;
 import org.teamcifo.tindergames.utils.Helpers;
 
-import java.util.HashSet;
-import java.util.Objects;
-import java.util.Set;
+import java.util.*;
 
 @AllArgsConstructor
 @Getter
 @Setter
 @ToString
-@EqualsAndHashCode
 // JPA annotations
 @Entity(name="User")
 @Table(name="USER_TABLE")
@@ -30,29 +30,39 @@ public class User {
     @Column(updatable = false, nullable = false)
     private String userId;
 
-    @OneToOne(cascade = CascadeType.ALL, mappedBy = "user")
-    @PrimaryKeyJoinColumn
-    private GamesCollection userGamesCollection;
+    @OneToMany(cascade = CascadeType.ALL)
+    @JoinTable(name = "game_statuses_mapping",
+            joinColumns = {@JoinColumn(name = "user_id", referencedColumnName = "userId")},
+            inverseJoinColumns = {@JoinColumn(name = "game_stats_id", referencedColumnName = "gameStatsId")})
+    @MapKeyJoinColumn(name = "game_id")
+    @ToString.Exclude
+    private Map<BoardGame, GameStats> userGamesCollection; // Keys are BoardGames
 
     @ManyToMany
     @JoinTable(name = "USER_GAMEPLAY",
             joinColumns = @JoinColumn(name = "USER_FK"),
             inverseJoinColumns = @JoinColumn(name = "GAMEPLAY_FK")
     )
+    @ToString.Exclude
     private Set<Gameplay> gameplays;
 
+    /*@ManyToMany(cascade={CascadeType.PERSIST, CascadeType.MERGE},
+            targetEntity=User.class)
+
+     */
     @ManyToMany
     @JoinTable(name = "USER_FRIENDS",
             joinColumns = @JoinColumn(name = "USER_FK"),
             inverseJoinColumns = @JoinColumn(name = "FRIEND_FK")
     )
+    @ToString.Exclude
     private Set<User> friends;
 
     public User() {
         this.userId = Helpers.generateUUID();
         this.gameplays = new HashSet<>();
         this.friends = new HashSet<>();
-        this.userGamesCollection = new GamesCollection(this);
+        this.userGamesCollection = new HashMap<>();
     }
 
     public User(String firstName, String lastName, String password, String email, String username) {
@@ -73,10 +83,29 @@ public class User {
     }
 
     public void addGameToCollection(BoardGame boardGame) {
-        this.getUserGamesCollection().addGameToCollection(boardGame);
+        this.userGamesCollection.putIfAbsent(boardGame, new GameStats());
     }
 
     public void deleteGameFromCollection(BoardGame boardGame) {
-        this.getUserGamesCollection().deleteGameFromCollection(boardGame);
+        if (this.hasGame(boardGame)) {
+            this.userGamesCollection.remove(boardGame);
+        }
+    }
+
+    public boolean hasGame(BoardGame boardGame) {
+        return this.userGamesCollection.containsKey(boardGame);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
+        User user = (User) o;
+        return getUserId() != null && Objects.equals(getUserId(), user.getUserId());
+    }
+
+    @Override
+    public int hashCode() {
+        return getClass().hashCode();
     }
 }
