+++ tindergames/src/test/java/org/teamcifo/utils/FakeDataGenerator.java
@@ -3,7 +3,6 @@ package org.teamcifo.utils;
 import com.github.javafaker.Faker;
 import org.teamcifo.tindergames.boardGameEntity.BoardGame;
 import org.teamcifo.tindergames.gamesCollectionEntity.GameStats;
-import org.teamcifo.tindergames.gamesCollectionEntity.GamesCollection;
 import org.teamcifo.tindergames.userEntity.User;
 
 import java.util.ArrayList;
